converted user score and ABV to float in DB

+++ lib/db/migrate/20200204154347_create_beer_table.rb
@@ -3,10 +3,10 @@ class CreateBeerTable < ActiveRecord::Migration[5.2]
     create_table :beers do |t|
       t.string :name
       t.string :brewery
-      t.integer :abv
+      t.float :abv
       t.string :style
       t.string :notes
-      t.integer :uscore
+      t.float :uscore
       t.integer :bascore
       t.string :img
     end
